+++ src/helpers/ConditionsHelper.php
@@ -101,6 +101,11 @@ class ConditionsHelper
                     'value' => $condition['value'],
                 ];
 
+                // Protect against empty conditions
+                if (!trim(ConditionsHelper::recursiveImplode('', $variables))) {
+                    continue;
+                }
+
                 $variables['field'] = str_replace(['{', '}'], ['', ''], $variables['field']);
 
                 // Check to see if this is a custom field, or an attribute on the submission
@@ -129,11 +134,6 @@ class ConditionsHelper
                     } else {
                         $variables['field'] = ArrayHelper::recursiveImplode(' ', $variables['field']);
                     }
-                }
-
-                // Protect against empty conditions
-                if (!trim(ArrayHelper::recursiveImplode('', $variables))) {
-                    continue;
                 }
 
                 // Create the rule for the evaluator - some rules need special syntax